Refactor/ff (#77)
* init

* cleanup api for RadiosityFast

* fix ruff

* remove patch_size

* move all classes to classes folder and rename files

* remove absorption

* fix init

* fix file name typo

* move form factor methods to form_factor/universal.py and ff helpers to geometry.py

* fix example

* fix all tests

* fix intent

* cleanup and rename ex "ff-helper" functions

* bring in visibility helper geometry methods

* update form_factor.py imports

* remove visibility helpers file

* change to brdf incoming and outgoing directions (adapted form graphics)

* fix universal form factor calls

* remove old universal form factor directory

* fix geometry module import

* fix test imports

* fix helper calls in universal ff methods

* ruff fixes

* make shoebox reciprocity check a tad faster

* rename universal methods file to avoid confusion with form_factor.universal

* fix visibility method calls in tests

* improve documentation of set_wall_brdf

* cleanup area calculation methods

* ruff fix

* ruff fix

* remove unused files

* rename to universial.py

* move scattering-related methods to RadiosityFast.py

* universial.py renaming fixes

* move energy exchange methods to RadiosityFast.py

* move scattering access and assignment to form_factor.py

* remove unused energy exchange algorithm

* move ff-specific geometry methods to dedicated file in form_factor/geometry.py

* bring general geometry methods to high-level geometry.py

* refactor: move remaining "fast" geometry methods to form_factor/integration.py

* fix tests

* remove unnecessary function

* include more generalized methods in integration.py

* split form_factor into universial.py and kang.py methods

* fix tests

* move receiver_energy.py to respective form_factor/kang.py or universial.py files

* fix tests

* move source_energy methods to respective form_factor files

* remove radiosity_fast directory

* fix __init__.py

* fix tests

* move directional scattering functions to radiosityfast

* fix directional scattering function calls after move

* ruff fixes

* remove unused form_factor.py

* rename/fix typo universial.py -> universal.py
no more risk of conflict

* make all geometry methods private

---------
