--- sparrowpy/radiosity_fast/__init__.py
@@ -1,7 +1,0 @@
-"""Module of the fast radiosity solver."""
-
-from . import geometry
-
-__all__ = [
-    'geometry',
-]
