--- sparrowpy/radiosity_fast/geometry.py
@@ -1,354 +1,0 @@
-"""Geometry functions for the radiosity fast solver."""
-try:
-    import numba
-    prange = numba.prange
-except ImportError:
-    numba = None
-    prange = range
-import numpy as np
-import sparrowpy.radiosity_fast.visibility_helpers as vh
-
-
-def get_scattering_data_receiver_index(
-        pos_i:np.ndarray, pos_j:np.ndarray,
-        receivers:np.ndarray, wall_id_i:np.ndarray,
-        ):
-    """Get scattering data depending on previous, current and next position.
-
-    Parameters
-    ----------
-    pos_i : np.ndarray
-        current position of shape (3)
-    pos_j : np.ndarray
-        next position of shape (3)
-    receivers : np.ndarray
-        receiver directions of all walls of shape (n_walls, n_receivers, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    n_patches = pos_i.shape[0] if pos_i.ndim > 1 else 1
-    receiver_idx = np.empty((n_patches), dtype=np.int64)
-
-    for i in range(n_patches):
-        difference_receiver = pos_i[i]-pos_j
-        difference_receiver /= np.linalg.norm(
-            difference_receiver)
-        receiver_idx[i] = np.argmin(np.sum(
-            (receivers[wall_id_i[i], :]-difference_receiver)**2, axis=-1),
-            axis=-1)
-
-
-    return receiver_idx
-
-
-
-def get_scattering_data(
-        pos_h:np.ndarray, pos_i:np.ndarray, pos_j:np.ndarray,
-        sources:np.ndarray, receivers:np.ndarray, wall_id_i:np.ndarray,
-        scattering:np.ndarray, scattering_index:np.ndarray):
-    """Get scattering data depending on previous, current and next position.
-
-    Parameters
-    ----------
-    pos_h : np.ndarray
-        previous position of shape (3)
-    pos_i : np.ndarray
-        current position of shape (3)
-    pos_j : np.ndarray
-        next position of shape (3)
-    sources : np.ndarray
-        source directions of all walls of shape (n_walls, n_sources, 3)
-    receivers : np.ndarray
-        receiver directions of all walls of shape (n_walls, n_receivers, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-    scattering : np.ndarray
-        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
-    scattering_index : np.ndarray
-        index of the scattering data of shape (n_walls)
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    difference_source = pos_h-pos_i
-    difference_receiver = pos_i-pos_j
-
-    difference_source /= np.linalg.norm(difference_source)
-    difference_receiver /= np.linalg.norm(difference_receiver)
-    source_idx = np.argmin(np.sum(
-        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
-    receiver_idx = np.argmin(np.sum(
-        (receivers[wall_id_i, :]-difference_receiver)**2, axis=-1))
-    return scattering[scattering_index[wall_id_i],
-        source_idx, receiver_idx, :]
-
-
-
-def get_scattering_data_source(
-        pos_h:np.ndarray, pos_i:np.ndarray,
-        sources:np.ndarray, wall_id_i:np.ndarray,
-        scattering:np.ndarray, scattering_index:np.ndarray):
-    """Get scattering data depending on previous, current position.
-
-    Parameters
-    ----------
-    pos_h : np.ndarray
-        previous position of shape (3)
-    pos_i : np.ndarray
-        current position of shape (3)
-    sources : np.ndarray
-        source directions of all walls of shape (n_walls, n_sources, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-    scattering : np.ndarray
-        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
-    scattering_index : np.ndarray
-        index of the scattering data of shape (n_walls)
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    difference_source = pos_h-pos_i
-    difference_source /= np.linalg.norm(difference_source)
-    source_idx = np.argmin(np.sum(
-        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
-    return scattering[scattering_index[wall_id_i], source_idx]
-
-
-def check_visibility(
-        patches_center:np.ndarray,
-        surf_normal:np.ndarray, surf_points:np.ndarray) -> np.ndarray:
-    """Check the visibility between patches.
-
-    Parameters
-    ----------
-    patches_center : np.ndarray
-        center points of all patches of shape (n_patches, 3)
-    surf_normal : np.ndarray
-        normal vectors of all patches of shape (n_patches, 3)
-    surf_points : np.ndarray
-        boundary points of possible blocking surfaces (n_surfaces,)
-
-    Returns
-    -------
-    visibility_matrix : np.ndarray
-        boolean matrix of shape (n_patches, n_patches) with True if patches
-        can see each other, otherwise false
-
-    """
-    n_patches = patches_center.shape[0]
-    visibility_matrix = np.empty((n_patches, n_patches), dtype=np.bool_)
-    visibility_matrix.fill(False)
-    indexes = []
-    for i_source in range(n_patches):
-        for i_receiver in range(n_patches):
-            if i_source < i_receiver:
-                indexes.append((i_source, i_receiver))
-                visibility_matrix[i_source,i_receiver]=True
-    indexes = np.array(indexes)
-    for i in prange(indexes.shape[0]):
-        i_source = indexes[i, 0]
-        i_receiver = indexes[i, 1]
-
-        surfid=0
-        while (visibility_matrix[i_source, i_receiver] and
-                                surfid!=len(surf_normal)):
-
-            visibility_matrix[i_source, i_receiver]=vh.basic_visibility(
-                                                            patches_center[i_source],
-                                                            patches_center[i_receiver],
-                                                            surf_points[surfid],
-                                                            surf_normal[surfid])
-
-            surfid+=1
-
-    return visibility_matrix
-
-
-def _create_patches(polygon_points:np.ndarray, max_size):
-    """Create patches from a polygon."""
-    size = np.empty(polygon_points.shape[1])
-    for i in range(polygon_points.shape[1]):
-        size[i] = polygon_points[:, i].max() - polygon_points[:, i].min()
-    patch_nums = np.array([int(n) for n in size/max_size])
-    real_size = size/patch_nums
-    if patch_nums[2] == 0:
-        x_idx = 0
-        y_idx = 1
-    if patch_nums[1] == 0:
-        x_idx = 0
-        y_idx = 2
-    if patch_nums[0] == 0:
-        x_idx = 1
-        y_idx = 2
-
-    x_min = np.min(polygon_points.T[x_idx])
-    y_min = np.min(polygon_points.T[y_idx])
-
-    n_patches = patch_nums[x_idx]*patch_nums[y_idx]
-    patches_points = np.empty((n_patches, 4, 3))
-    i = 0
-    for i_x in range(patch_nums[x_idx]):
-        for i_y in range(patch_nums[y_idx]):
-            points = polygon_points.copy()
-            points[0, x_idx] = x_min + i_x * real_size[x_idx]
-            points[0, y_idx] = y_min + i_y * real_size[y_idx]
-            points[1, x_idx] = x_min + (i_x+1) * real_size[x_idx]
-            points[1, y_idx] = y_min + i_y * real_size[y_idx]
-            points[3, x_idx] = x_min + i_x * real_size[x_idx]
-            points[3, y_idx] = y_min + (i_y+1) * real_size[y_idx]
-            points[2, x_idx] = x_min + (i_x+1) * real_size[x_idx]
-            points[2, y_idx] = y_min + (i_y+1) * real_size[y_idx]
-            patches_points[i] = points
-            i += 1
-
-    return patches_points
-
-
-def _calculate_center(points):
-    return np.sum(points, axis=-2) / points.shape[-2]
-
-
-def _calculate_size(points):
-    vec1 = points[..., 0, :]-points[..., 1, :]
-    vec2 = points[..., 1, :]-points[..., 2, :]
-    return np.abs(vec1-vec2)
-
-
-def _calculate_area(points):
-    area = np.zeros(points.shape[0])
-
-    for i in prange(points.shape[0]):
-        for tri in range(points.shape[1]-2):
-            area[i] +=  .5 * np.linalg.norm(
-                np.cross(
-                    points[i, tri+1,:] - points[i, 0,:],
-                    points[i, tri+2,:]-points[i, 0,:],
-                    ),
-                )
-
-    return area
-
-
-def process_patches(
-        polygon_points_array: np.ndarray,
-        walls_normal: np.ndarray,
-        patch_size:float, n_walls:int):
-    """Process the patches.
-
-    Parameters
-    ----------
-    polygon_points_array : np.ndarray
-        points of the polygon of shape (n_walls, 4, 3)
-    walls_normal : np.ndarray
-        wall normal of shape (n_walls, 3)
-    patch_size : float
-        maximal patch size in meters of shape (n_walls, 3).
-    n_walls : int
-        number of walls
-
-    Returns
-    -------
-    patches_points : np.ndarray
-        points of all patches of shape (n_patches, 4, 3)
-    patches_normal : np.ndarray
-        normal of all patches of shape (n_patches, 3)
-    n_patches : int
-        number of patches
-
-    """
-    n_patches = 0
-    n_walls = polygon_points_array.shape[0]
-    for i in range(n_walls):
-        n_patches += total_number_of_patches(
-            polygon_points_array[i, :, :], patch_size)
-    patches_points = np.empty((n_patches, 4, 3))
-    patch_to_wall_ids = np.empty((n_patches), dtype=np.int64)
-    patches_per_wall = np.empty((n_walls), dtype=np.int64)
-
-    for i in range(n_walls):
-        polygon_points = polygon_points_array[i, :]
-        patches_points_wall = _create_patches(
-            polygon_points, patch_size)
-        patches_per_wall[i] = patches_points_wall.shape[0]
-        j_start = (np.sum(patches_per_wall[:i])) if i > 0 else 0
-        j_end = np.sum(patches_per_wall[:i+1])
-        patch_to_wall_ids[j_start:j_end] = i
-        patches_points[j_start:j_end, :, :] = patches_points_wall
-    n_patches = patches_points.shape[0]
-
-    # calculate patch information
-    patches_normal = walls_normal[patch_to_wall_ids, :]
-    return (patches_points, patches_normal, n_patches, patch_to_wall_ids)
-
-
-def total_number_of_patches(polygon_points:np.ndarray, max_size: float):
-    """Calculate the total number of patches.
-
-    Parameters
-    ----------
-    polygon_points : np.ndarray
-        points of the polygon of shape (4, 3)
-    max_size : float
-        maximal patch size in meters
-
-    Returns
-    -------
-    n_patches : int
-        number of patches
-
-    """
-    size = np.empty(polygon_points.shape[1])
-    for i in range(polygon_points.shape[1]):
-        size[i] = polygon_points[:, i].max() - polygon_points[:, i].min()
-    patch_nums = np.array([int(n) for n in size/max_size])
-    if patch_nums[2] == 0:
-        x_idx = 0
-        y_idx = 1
-    if patch_nums[1] == 0:
-        x_idx = 0
-        y_idx = 2
-    if patch_nums[0] == 0:
-        x_idx = 1
-        y_idx = 2
-
-    return patch_nums[x_idx]*patch_nums[y_idx]
-
-def calculate_normals(points: np.ndarray):
-    """Calculate normals of plane defined by 3 or more input points."""
-
-    normals = np.empty((points.shape[0],3))
-
-    for i in numba.prange(points.shape[0]):
-        normals[i]=np.cross(points[i][1]-points[i][0],points[i][2]-points[i][0])
-        normals[i]/=np.linalg.norm(normals[i])
-
-    return normals
-
-
-if numba is not None:
-    get_scattering_data_receiver_index = numba.njit()(
-        get_scattering_data_receiver_index)
-    total_number_of_patches = numba.njit()(total_number_of_patches)
-    process_patches = numba.njit()(process_patches)
-    _calculate_area = numba.njit()(_calculate_area)
-    _calculate_center = numba.njit()(_calculate_center)
-    _calculate_size = numba.njit()(_calculate_size)
-    _create_patches = numba.njit()(_create_patches)
-    get_scattering_data = numba.njit()(get_scattering_data)
-    get_scattering_data_source = numba.njit()(get_scattering_data_source)
-    check_visibility = numba.njit(parallel=True)(check_visibility)
-    calculate_normals = numba.njit()(calculate_normals)
-
