--- sparrowpy/radiosity_fast/universal_ff/__init__.py
@@ -1,1 +1,0 @@
-"""Helper for form factor calculation."""
