+++ tests/test_visibility.py
@@ -1,8 +1,7 @@
 import numpy.testing as npt
 import pytest
-import sparrowpy.radiosity_fast.visibility_helpers as vh
 import numpy as np
-import sparrowpy.radiosity_fast.geometry as geom
+import sparrowpy.geometry as geom
 bpy = pytest.importorskip("bpy")
 import sparrowpy.utils.blender as bh  # noqa: E402
 import matplotlib.pyplot as plt  # noqa: E402
@@ -15,7 +14,7 @@ import matplotlib.pyplot as plt  # noqa: E402
 def test_point_plane_projection(origin: np.ndarray, point: np.ndarray,
                                 plpt: np.ndarray, pln: np.ndarray, solution):
     """Ensure correct projection of rays into plane."""
-    out = vh.project_to_plane(origin, point, plpt, pln)
+    out = geom._project_to_plane(origin, point, plpt, pln)
 
     npt.assert_array_equal(solution,out)
 
@@ -31,7 +30,9 @@ def test_point_plane_projection(origin: np.ndarray, point: np.ndarray,
 @pytest.mark.parametrize("pln", [np.array([0.,0.,1.])])
 def test_point_in_polygon(point, plpt, pln):
     """Ensure correct projection of rays into plane."""
-    out = vh.point_in_polygon(point3d=point, polygon3d=plpt, plane_normal=pln)
+    out = geom._point_in_polygon(point3d=point,
+                                 polygon3d=plpt,
+                                 plane_normal=pln)
 
     if abs(point[0]) > 1. or abs(point[1]) > 1:
         solution = False
@@ -56,7 +57,7 @@ def test_point_in_polygon(point, plpt, pln):
     ])
 def test_basic_visibility(point, origin, plpt, pln):
     """Test basic_visibility function."""
-    out = vh.basic_visibility(eval_point=point,vis_point=origin,
+    out = geom._basic_visibility(eval_point=point,vis_point=origin,
                               surf_points=plpt,surf_normal=pln)
 
     if (abs(point[0])/(-point[2]+1) > 1. or
@@ -125,7 +126,7 @@ def test_vis_matrix_assembly(model):
                                             -m["verts"][m["conn"][i]][0])
             surfs_normals[i]/=np.linalg.norm(surfs_normals[i])
 
-        vis_matrix = geom.check_visibility(patches_center=patches_centers,
+        vis_matrix = geom._check_visibility(patches_center=patches_centers,
                                            surf_normal=surfs_normals,
                                            surf_points=surfs_points)
